Rca/python combined syzygy tool
This is a combined SZYGY DNA tool. It ports the functions of the dna-writer and sequence-writer, and adds the ability to program a peripheral with firmware, DNA and sequence data all in one step. With it you can fully program a peripheral with just an Atmel programmer and don't need any I2C host (such as the Brain-1) or additional scripts/steps. The current boards serial and JSON DNA config file are taken as arguments.